Move logic to agent + frontend
