+++ coderd/coderd.go
@@ -201,7 +201,7 @@ func New(options *Options) *API {
 			httpmw.ExtractUserParam(api.Database),
 			// Extracts the <workspace.agent> from the url
 			httpmw.ExtractWorkspaceAndAgentParam(api.Database),
-			httpmw.ActivityBumpWorkspace(api.Logger, api.Database),
+			// httpmw.ActivityBumpWorkspace(api.Logger, api.Database),
 		)
 		r.HandleFunc("/*", api.workspaceAppsProxyPath)
 	}
@@ -422,7 +422,6 @@ func New(options *Options) *API {
 					apiKeyMiddleware,
 					httpmw.ExtractWorkspaceAgentParam(options.Database),
 					httpmw.ExtractWorkspaceParam(options.Database),
-					httpmw.ActivityBumpWorkspace(api.Logger, options.Database),
 				)
 				r.Get("/", api.workspaceAgent)
 				r.Get("/pty", api.workspaceAgentPTY)
