+++ site/src/components/WorkspaceScheduleForm/WorkspaceScheduleForm.tsx
@@ -55,7 +55,8 @@ export const Language = {
   timezoneLabel: "Timezone",
   ttlLabel: "Time until shutdown (hours)",
   ttlCausesShutdownHelperText: "Your workspace will shut down",
-  ttlCausesShutdownAfterStart: "after its next start",
+  ttlCausesShutdownAfterStart: `after its next start. We delay shutdown by an
+hour whenever we detect activity`,
   ttlCausesNoShutdownHelperText: "Your workspace will not automatically shut down.",
   formTitle: "Workspace schedule",
   startSection: "Start",
